Chore: added toLong to recordLimit and byteLimit in batchingSettings

+++ 2-collectors/scala-stream-collector/pubsub/src/main/scala/com.snowplowanalytics.snowplow.collectors.scalastream/sinks/GooglePubSubSink.scala
@@ -73,8 +73,8 @@ object GooglePubSubSink {
   private def batchingSettings(bufferConfig: BufferConfig): BatchingSettings =
     BatchingSettings
       .newBuilder()
-      .setElementCountThreshold(bufferConfig.recordLimit)
-      .setRequestByteThreshold(bufferConfig.byteLimit)
+      .setElementCountThreshold(bufferConfig.recordLimit.toLong)
+      .setRequestByteThreshold(bufferConfig.byteLimit.toLong)
       .setDelayThreshold(Duration.ofMillis(bufferConfig.timeLimit))
       .build()
